// frontend/src/app/swagger/model/categoryReadDto.ts
/**
 * S2Retro.Api
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CategoryValueReadDto } from './categoryValueReadDto';


export interface CategoryReadDto { 
    id?: string;
    name?: string | null;
    values?: Array<CategoryValueReadDto> | null;
}